refactor applicationcontroller and separate jwt feature into a module
